I'm done with my changes.

+++ Gustav Fransson/Assets/Uppgift3.cs
@@ -1,19 +1,22 @@
 ﻿using UnityEngine;
 using System.Collections;
 
-public class Uppgift3 : MonoBehaviour {
-    public float a;
-    public float b;
-    public float c;
-	// I create variables
+public class Uppgift3 : MonoBehaviour 
+{
+    // Created 2 variables
+    public int a;
+    public int b;
+    public int c;
+	
 	void Start () {
         
         
         
 	}
 	
-	// A method that adds a and b.
-	public int AddNumber (int a, int b) {
+	// A method that adds a and b and then returning their values.
+	public int AddNumber (int a, int b) 
+    {
 	return a + b;
 	}
 }
